+++ atomicshop/etws/traces/trace_dns.py
@@ -1,18 +1,31 @@
-from . import trace
-from .. import dns
-from ..wrappers.psutilw import psutilw
-from ..basics import dicts
-from ..process_poller import ProcessPollerPool
-from ..print_api import print_api
+import time
+
+from .. import trace, const
+from ...wrappers.psutilw import psutilw
+from ...basics import dicts
+from ...process_poller import ProcessPollerPool
+from ...print_api import print_api
+from ... import dns
 
 
-class DnsTrace:
+ETW_DEFAULT_SESSION_NAME: str = 'AtomicShopDnsTrace'
+
+PROVIDER_NAME: str = const.ETW_DNS['provider_name']
+PROVIDER_GUID: str = const.ETW_DNS['provider_guid']
+REQUEST_RESP_EVENT_ID: int = const.ETW_DNS['event_ids']['dns_request_response']
+
+WAIT_FOR_PROCESS_POLLER_PID_SECONDS: int = 3
+WAIT_FOR_PROCESS_POLLER_PID_COUNTS: int = WAIT_FOR_PROCESS_POLLER_PID_SECONDS * 10
+
+
+class DnsRequestResponseTrace:
     def __init__(
             self,
-            enable_process_poller: bool = False,
             attrs: list = None,
             session_name: str = None,
-            close_existing_session_name: bool = True
+            close_existing_session_name: bool = True,
+            enable_process_poller: bool = False,
+            process_poller_etw_session_name: str = None
     ):
         """
         DnsTrace class use to trace DNS events from Windows Event Tracing for EventId 3008.
@@ -28,6 +41,7 @@ class DnsTrace:
             False: if ETW session with 'session_name' exists, you will be notified and the new session will not be
                 created. Instead, the existing session will be used. If there is a buffer from the previous session,
                 you will get the events from the buffer.
+        :param process_poller_etw_session_name: The name of the ETW session for tracing process creation.
 
         -------------------------------------------------
 
@@ -35,7 +49,13 @@ class DnsTrace:
             from atomicshop.etw import dns_trace
 
 
-            dns_trace_w = dns_trace.DnsTrace(enable_process_poller=True, attrs=['pid', 'name', 'cmdline', 'domain', 'query_type'])
+            dns_trace_w = dns_trace.DnsTrace(
+                attrs=['pid', 'name', 'cmdline', 'domain', 'query_type'],
+                session_name='MyDnsTrace',
+                close_existing_session_name=True,
+                enable_process_poller=True,
+                process_poller_etw_session_name='MyProcessTrace'
+            )
             dns_trace_w.start()
             while True:
                 dns_dict = dns_trace_w.emit()
@@ -46,16 +66,21 @@ class DnsTrace:
         self.enable_process_poller = enable_process_poller
         self.attrs = attrs
 
+        if not session_name:
+            session_name = ETW_DEFAULT_SESSION_NAME
+
         self.event_trace = trace.EventTrace(
-            providers=[(dns.ETW_DNS_INFO['provider_name'], dns.ETW_DNS_INFO['provider_guid'])],
+            providers=[(PROVIDER_NAME, PROVIDER_GUID)],
             # lambda x: self.event_queue.put(x),
-            event_id_filters=[dns.ETW_DNS_INFO['event_id']],
+            event_id_filters=[REQUEST_RESP_EVENT_ID],
             session_name=session_name,
             close_existing_session_name=close_existing_session_name
         )
 
         if self.enable_process_poller:
-            self.process_poller = ProcessPollerPool(store_cycles=200, operation='process', poller_method='process_dll')
+            self.process_poller = ProcessPollerPool(
+                operation='process', poller_method='sysmon_etw',
+                sysmon_etw_session_name=process_poller_etw_session_name)
 
     def start(self):
         if self.enable_process_poller:
@@ -82,6 +107,11 @@ class DnsTrace:
         :return: Dictionary with the event data.
         """
 
+        # Get the processes first, since we need the process name and command line.
+        # If they're not ready, we will get just pids from DNS tracing.
+        if self.enable_process_poller:
+            self._get_processes_from_poller()
+
         event = self.event_trace.emit()
 
         event_dict: dict = {
@@ -91,12 +121,6 @@ class DnsTrace:
             'query_type_id': str(event[1]['QueryType']),
             'query_type': dns.TYPES_DICT[str(event[1]['QueryType'])]
         }
-
-        # # Get process name only from psutil, just in case.
-        # try:
-        #     process_name = psutilw.get_process_name_by_pid(event_dict['pid'])
-        # except psutil.NoSuchProcess:
-        #     process_name = str(event_dict['pid'])
 
         # Defining list if ips and other answers, which aren't IPs.
         list_of_ips = list()
@@ -135,10 +159,19 @@ class DnsTrace:
             event_dict['status'] = 'Error'
 
         if self.enable_process_poller:
-            processes = self.process_poller.processes
+            processes = self.process_poller.get_processes()
+            if event_dict['pid'] not in processes:
+                counter = 0
+                while counter < WAIT_FOR_PROCESS_POLLER_PID_COUNTS:
+                    processes = self.process_poller.get_processes()
+                    if event_dict['pid'] not in processes:
+                        time.sleep(0.1)
+                        counter += 1
+                    else:
+                        break
 
-            if isinstance(processes, BaseException):
-                raise processes
+                if counter == WAIT_FOR_PROCESS_POLLER_PID_COUNTS:
+                    print_api(f"Error: Couldn't get the process name for PID: {event_dict['pid']}.", color='red')
 
             event_dict = psutilw.cross_single_connection_with_processes(event_dict, processes)
             # If it was impossible to get the process name from the process poller, get it from psutil.
@@ -150,3 +183,13 @@ class DnsTrace:
                 event_dict, self.attrs, skip_keys_not_in_list=True)
 
         return event_dict
+
+    def _get_processes_from_poller(self):
+        processes: dict = {}
+        while not processes:
+            processes = self.process_poller.get_processes()
+
+            if isinstance(processes, BaseException):
+                raise processes
+
+        return processes